✨ Feat: 인증샷 업로드

+++ backend/src/main/java/com/a101/ecofarming/proof/dto/response/ProofUploadResponseDto.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class ProofUploadResponseDto {
     private Integer proofId;
-    private Integer progress;
+    private Byte progress;
 }
